requirements.txt
Flask==2.3.3
APScheduler==3.10.4
python-nmap==0.7.1
SQLAlchemy==2.0.21
python-dotenv==1.0.0
requests==2.31.0
mysqlclient==2.2.0
Flask-SQLAlchemy==3.1.1
